+++ DSCResources/MSFT_xSQLServerEndpointState/MSFT_xSQLServerEndpointState.psm1
@@ -24,9 +24,9 @@ function Get-TargetResource
         [System.String]
         $InstanceName,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.String]
-        $NodeName,
+        $NodeName = $env:COMPUTERNAME,
 
         [Parameter(Mandatory = $true)]
         [System.String]
@@ -87,9 +87,9 @@ function Set-TargetResource
         [System.String]
         $InstanceName,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.String]
-        $NodeName,
+        $NodeName = $env:COMPUTERNAME,
 
         [Parameter(Mandatory = $true)]
         [System.String]
@@ -162,9 +162,9 @@ function Test-TargetResource
         [System.String]
         $InstanceName,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.String]
-        $NodeName,
+        $NodeName = $env:COMPUTERNAME,
 
         [Parameter(Mandatory = $true)]
         [System.String]