+++ CHANGELOG.md
@@ -104,10 +104,12 @@
   - Added unit tests (issue #289)
   - Added examples.
 - Changes to xSQLServerEndpointState
-  - Cleaned up code, removed `SupportsShouldProcess` and fixed PSSA rules warnings (issue #258 and issue #230).
-  - Now the defult value for the parameter `State` is 'Started'.
+  - Cleaned up code, removed SupportsShouldProcess and fixed PSSA rules warnings (issue #258 and issue #230).
+  - Now the defult value for the parameter State is 'Started'.
   - Updated README.md with a description for the resources and revised the parameter descriptions.
   - Removed dependency of SQLPS provider (issue #481).
+  - The parameter NodeName is no longer mandatory and has now the default value of $env:COMPUTERNAME.
+  - The parameter Name is now a key so it is now possible to change the state on more than one endpoint on the same instance. _Note: The resource still only supports Database Mirror endpoints at this time._
 
 ## 6.0.0.0